+++ src/strife/r_plane.cpp
@@ -153,7 +153,7 @@ R_MapPlane
 	ds_colormap = fixedcolormap;
     else
     {
-	index = distance >> LIGHTZSHIFT;
+	index = static_cast<unsigned int>(distance >> LIGHTZSHIFT);
 	
 	if (index >= MAXLIGHTZ )
 	    index = MAXLIGHTZ-1;
